Versión 1.1
Añadida función de cuenta total

+++ src/eb/controller/Controller.java
@@ -40,7 +40,7 @@ public class Controller {
                     for(int row=0; row<rowSize; row++){
                         if(rSet.next()){
                             table[row][0] = rSet.getString("Platillo");
-                            table[row][1] = "" +rSet.getFloat("Costo");
+                            table[row][1] = "" +rSet.getInt("Costo");
                         }
                     }
                 query.closeDBQuery();
